+++ testing/upload_validation.ts
@@ -92,6 +92,15 @@ import * as z from 'zod';
  */
 export const PACKS_VALID_COLUMN_FORMAT_MATCHER_REGEX = /^\/(.*)\/([a-z]+)?$/;
 
+export const Limits = {
+  BuildingBlockCountPerType: 100,
+  BuildingBlockName: 50,
+  BuildingBlockDescription: 250,
+  ColumnMatcherRegex: 300,
+  NumColumnMatchersPerFormat: 10,
+  NetworkDomainUrl: 253,
+}
+
 enum CustomErrorCode {
   NonMatchingDiscriminant = 'nonMatchingDiscriminant',
 }
@@ -466,7 +475,7 @@ const variousSupportedAuthenticationValidators = Object.entries(defaultAuthentic
 const primitiveUnion = z.union([z.number(), z.string(), z.boolean(), z.date()]);
 
 const paramDefValidator = zodCompleteObject<ParamDef<any>>({
-  name: z.string(),
+  name: z.string().max(Limits.BuildingBlockName),
   type: z
     .union([
       z.nativeEnum(Type),
@@ -484,7 +493,7 @@ const paramDefValidator = zodCompleteObject<ParamDef<any>>({
         message: 'Object parameters are not currently supported.',
       },
     ),
-  description: z.string(),
+  description: z.string().max(Limits.BuildingBlockDescription),
   optional: z.boolean().optional(),
   autocomplete: z.unknown().optional(),
   defaultValue: z.unknown().optional(),
@@ -495,8 +504,8 @@ const commonPackFormulaSchema = {
   // It would be preferable to use validateFormulaName here, but we have to exempt legacy packs with sync tables
   // whose getter names violate the validator, and those exemptions require the pack id, so this has to be
   // done as a superRefine on the top-level object that also contains the pack id.
-  name: z.string(),
-  description: z.string(),
+  name: z.string().max(Limits.BuildingBlockName),
+  description: z.string().max(Limits.BuildingBlockDescription),
   examples: z
     .array(
       z.object({
@@ -903,42 +912,42 @@ const genericObjectSchema: z.ZodTypeAny = z.lazy(() =>
       });
     })
     .superRefine((data, context) => {
-      const schemaHelper = objectSchemaHelper(data as GenericObjectSchema);
+      const schema = data as GenericObjectSchema;
 
       const validateTitleProperty = () => {
-        if (schemaHelper.titleProperty) {
-          if (!(schemaHelper.titleProperty in schemaHelper.properties)) {
+        if (schema.titleProperty) {
+          if (!(schema.titleProperty in schema.properties)) {
             context.addIssue({
               code: z.ZodIssueCode.custom,
               path: ['titleProperty'],
-              message: `The "titleProperty" field name "${schemaHelper.titleProperty}" does not exist in the "properties" object.`,
+              message: `The "titleProperty" field name "${schema.titleProperty}" does not exist in the "properties" object.`,
             });
             return;
           }
 
-          const titlePropertySchema = schemaHelper.properties[schemaHelper.titleProperty];
-          if (titlePropertySchema.type !== ValueType.String) {
+          const titlePropertySchema = schema.properties[schema.titleProperty];
+          if (![ValueType.String, ValueType.Object].includes(titlePropertySchema.type)) {
             context.addIssue({
               code: z.ZodIssueCode.custom,
               path: ['titleProperty'],
-              message: `The "titleProperty" field name "${schemaHelper.titleProperty}" must refer to a "ValueType.String" property.`,
+              message: `The "titleProperty" field name "${schema.titleProperty}" must refer to a "ValueType.String" or "ValueType.Object" property.`,
             });
             return;
           }
         }
       };
       const validateImageProperty = () => {
-        if (schemaHelper.imageProperty) {
-          if (!(schemaHelper.imageProperty in schemaHelper.properties)) {
+        if (schema.imageProperty) {
+          if (!(schema.imageProperty in schema.properties)) {
             context.addIssue({
               code: z.ZodIssueCode.custom,
               path: ['imageProperty'],
-              message: `The "imageProperty" field name "${schemaHelper.imageProperty}" does not exist in the "properties" object.`,
+              message: `The "imageProperty" field name "${schema.imageProperty}" does not exist in the "properties" object.`,
             });
             return;
           }
 
-          const imagePropertySchema = schemaHelper.properties[schemaHelper.imageProperty];
+          const imagePropertySchema = schema.properties[schema.imageProperty];
           if (
             imagePropertySchema.type !== ValueType.String ||
             ![ValueHintType.ImageAttachment, ValueHintType.ImageReference].includes(
@@ -948,24 +957,24 @@ const genericObjectSchema: z.ZodTypeAny = z.lazy(() =>
             context.addIssue({
               code: z.ZodIssueCode.custom,
               path: ['imageProperty'],
-              message: `The "imageProperty" field name "${schemaHelper.imageProperty}" must refer to a "ValueType.String" property with a "ValueHintType.ImageAttachment" or "ValueHintType.ImageReference" "codaType".`,
+              message: `The "imageProperty" field name "${schema.imageProperty}" must refer to a "ValueType.String" property with a "ValueHintType.ImageAttachment" or "ValueHintType.ImageReference" "codaType".`,
             });
             return;
           }
         }
       };
       const validateDescriptionProperty = () => {
-        if (schemaHelper.descriptionProperty) {
-          if (!(schemaHelper.descriptionProperty in schemaHelper.properties)) {
+        if (schema.descriptionProperty) {
+          if (!(schema.descriptionProperty in schema.properties)) {
             context.addIssue({
               code: z.ZodIssueCode.custom,
               path: ['descriptionProperty'],
-              message: `The "descriptionProperty" field name "${schemaHelper.descriptionProperty}" does not exist in the "properties" object.`,
+              message: `The "descriptionProperty" field name "${schema.descriptionProperty}" does not exist in the "properties" object.`,
             });
             return;
           }
 
-          const descriptionPropertySchema = schemaHelper.properties[schemaHelper.descriptionProperty];
+          const descriptionPropertySchema = schema.properties[schema.descriptionProperty];
           if (
             descriptionPropertySchema.type !== ValueType.String &&
             (descriptionPropertySchema.type !== ValueType.Array ||
@@ -974,29 +983,29 @@ const genericObjectSchema: z.ZodTypeAny = z.lazy(() =>
             context.addIssue({
               code: z.ZodIssueCode.custom,
               path: ['descriptionProperty'],
-              message: `The "descriptionProperty" field name "${schemaHelper.descriptionProperty}" must refer to a "ValueType.String" property or array of strings.`,
+              message: `The "descriptionProperty" field name "${schema.descriptionProperty}" must refer to a "ValueType.String" property or array of strings.`,
             });
             return;
           }
         }
       };
       const validateLinkProperty = () => {
-        if (schemaHelper.linkProperty) {
-          if (!(schemaHelper.linkProperty in schemaHelper.properties)) {
+        if (schema.linkProperty) {
+          if (!(schema.linkProperty in schema.properties)) {
             context.addIssue({
               code: z.ZodIssueCode.custom,
               path: ['linkProperty'],
-              message: `The "linkProperty" field name "${schemaHelper.linkProperty}" does not exist in the "properties" object.`,
+              message: `The "linkProperty" field name "${schema.linkProperty}" does not exist in the "properties" object.`,
             });
             return;
           }
 
-          const linkPropertySchema = schemaHelper.properties[schemaHelper.linkProperty];
+          const linkPropertySchema = schema.properties[schema.linkProperty];
           if (linkPropertySchema.type !== ValueType.String || linkPropertySchema.codaType !== ValueHintType.Url) {
             context.addIssue({
               code: z.ZodIssueCode.custom,
               path: ['linkProperty'],
-              message: `The "linkProperty" field name "${schemaHelper.linkProperty}" must refer to a "ValueType.String" property with a "ValueHintType.Url" "codaType".`,
+              message: `The "linkProperty" field name "${schema.linkProperty}" must refer to a "ValueType.String" property with a "ValueHintType.Url" "codaType".`,
             });
             return;
           }
@@ -1007,8 +1016,8 @@ const genericObjectSchema: z.ZodTypeAny = z.lazy(() =>
       validateLinkProperty();
       validateImageProperty();
       validateDescriptionProperty();
-      (schemaHelper.subtitleProperties || []).forEach((f, i) => {
-        if (!(f in schemaHelper.properties)) {
+      (schema.subtitleProperties || []).forEach((f, i) => {
+        if (!(f in schema.properties)) {
           context.addIssue({
             code: z.ZodIssueCode.custom,
             path: ['subtitleProperties', i],
@@ -1017,7 +1026,7 @@ const genericObjectSchema: z.ZodTypeAny = z.lazy(() =>
           return;
         }
 
-        const subtitlePropertySchema = schemaHelper.properties[f];
+        const subtitlePropertySchema = schema.properties[f];
         if (
           'codaType' in subtitlePropertySchema &&
           subtitlePropertySchema.codaType &&
@@ -1076,13 +1085,14 @@ const formulaMetadataSchema = z
   });
 
 const formatMetadataSchema = zodCompleteObject<PackFormatMetadata>({
-  name: z.string(),
+  name: z.string().max(Limits.BuildingBlockName),
   formulaNamespace: z.string().optional(), // Will be removed once we deprecate namespace objects.
   formulaName: z.string(),
   hasNoConnection: z.boolean().optional(),
   instructions: z.string().optional(),
   placeholder: z.string().optional(),
-  matchers: z.array(z.string().refine(validateFormatMatcher)),
+  matchers: z.array(
+    z.string().max(Limits.ColumnMatcherRegex).refine(validateFormatMatcher)).max(Limits.NumColumnMatchersPerFormat),
 });
 
 const syncFormulaSchema = zodCompleteObject<Omit<SyncFormula<any, any, ParamDefs, ObjectSchema<any, any>>, 'execute'>>({
@@ -1093,8 +1103,8 @@ const syncFormulaSchema = zodCompleteObject<Omit<SyncFormula<any, any, ParamDefs
 });
 
 const baseSyncTableSchema = {
-  name: z.string().nonempty(),
-  description: z.string().optional(),
+  name: z.string().nonempty().max(Limits.BuildingBlockName),
+  description: z.string().max(Limits.BuildingBlockDescription).optional(),
   schema: genericObjectSchema,
   getter: syncFormulaSchema,
   entityName: z.string().optional(),
@@ -1103,6 +1113,7 @@ const baseSyncTableSchema = {
   identityName: z
     .string()
     .min(1)
+    .max(Limits.BuildingBlockName)
     .optional()
     .refine(
       val => !val || !SystemColumnNames.includes(val),
@@ -1163,6 +1174,7 @@ const unrefinedPackVersionMetadataSchema = zodCompleteObject<PackVersionMetadata
     .array(
       z
         .string()
+        .max(Limits.NetworkDomainUrl)
         .refine(domain => !(domain.startsWith('http:') || domain.startsWith('https:') || domain.indexOf('/') >= 0), {
           message: 'Invalid network domain. Instead of "https://www.example.com", just specify "example.com".',
         }),
@@ -1172,10 +1184,11 @@ const unrefinedPackVersionMetadataSchema = zodCompleteObject<PackVersionMetadata
     message: 'Formula namespaces can only contain alphanumeric characters and underscores.',
   }),
   systemConnectionAuthentication: z.union(zodUnionInput(systemAuthenticationValidators)).optional(),
-  formulas: z.array(formulaMetadataSchema).optional().default([]),
-  formats: z.array(formatMetadataSchema).optional().default([]),
+  formulas: z.array(formulaMetadataSchema).max(Limits.BuildingBlockCountPerType).optional().default([]),
+  formats: z.array(formatMetadataSchema).max(Limits.BuildingBlockCountPerType).optional().default([]),
   syncTables: z
     .array(syncTableSchema)
+    .max(Limits.BuildingBlockCountPerType)
     .optional()
     .default([])
     .superRefine((data, context) => {